[TASK] Update ckeditor to 4.6.1
New Features:
* The callback parameter in the CKEDITOR.ajax.post method became optional.

Fixed Issues:
* [Blink, WebKit] Fixed: Cannot select all editor content when a widget or
a non-editable element is the first or last element of the content. Also
fixes this issue in the Select All plugin.
* [Blink, WebKit, IE8] Fixed: Browser hangs when a table is inserted in
the place of a selected list with an empty last item.
* Fixed: Improved the Color Button plugin which will now normalize the CSS
background property if it only contains a color value. This fixes missing
background colors when using Paste from Word.
+ [Blink, WebKit] Fixed: Error thrown occasionally by an uninitialized
editable for multiple CKEditor instances on the same page.

Resolves: #78966
Releases: master
Change-Id: I2e711c363dcc8a4a00f7e88828105de987968da4
Reviewed-on: https://review.typo3.org/50939
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>