[TASK] Always keep <a> instead of <link> tags in RTE transformations
The RTE uses <a> tags for handling links, a transformation
that replaces <link> tags to <a> tags is updated to use the
new t3:// linking syntax instead of hard-coding index.php?id=23
or trying to manually resolve the TypoLink functionality
for doing links.

This also means that the RTE now deals with t3:// links,
and HtmlArea is adapted as well.

When saving the links in the database again, the links
are now not transformed into <link> tags again but kept
as proper <a> tags.

Resolves: #79291
Releases: master
Change-Id: Idc002138e1283d6962666dcd44ba3039c464d6ca
Reviewed-on: https://review.typo3.org/49441
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>