[BUGFIX] Refactor CKEditor to be inline with the AbstractLinkBrowser
The internals of the LinkBrowser and LinkHandler depend on all the
field specific info to be in the P parameter.

Further the DatabaseBrowser used by the RecordBrowser fetches the table
parameter form global url query scope what collides with the params set
by the CKEditor RichTextElement.

To prevent the parameter collision the params set by the RichTextElement
are moved to the P prefix what is also more inline with the other
LinkBrowser implementations.

This patch also resolves the lasts issues related to the new 
recordHandler + CKEditor.

Resolves: #79645
Releases: master
Change-Id: Id7cab0f0a012c5de11a2038e1d579d87599d47cd
Reviewed-on: https://review.typo3.org/51760
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>