[TASK] CKEditor: Add quicktable plugin / easier table configuration
In order to reduce the dialog popups and ease the pain with tables
within CKEditor, quicktable (https://github.com/ufdada/quicktable)
is added which complements the basic "table" plugin.

The quicktable plugin is modified to add only unstyled tables
instead of the default tables/borders.

Additionally, some styles for tables are added to the default contents.css

Resolves: #80131
Releases: master
Change-Id: I06b619fa89e21cb43ea79ebc8a67238db8d58c66
Reviewed-on: https://review.typo3.org/51945
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>