[BUGFIX] CKEditor: UI language should respect BE user language
If a backend user has selected a backend language,
CKEditor is not respecting this option, instead calculating
this based on the current browsers' language.

By default, CKeditor checks if the config option "language"
is set and if this is a valid language key, and translation is available
if not, it checks for the browser settings, if nothing matching found
it falls back to to the config option "defaultLanguage".

"defaultLanguage" should be set to "en", as the TYPO3 Backend
does as well.

The config option "language" should be calculated based on the
Backend Users' settings.

Resolves: #80133
Releases: master
Change-Id: I0f095ab7147871c2a4e8c8ea379754749bc1ba98
Reviewed-on: https://review.typo3.org/51947
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>