[BUGFIX] Skip CKEditor registration on CLI
Made sure the CKEditor RequireJS paths are not calculated when running
in a CLI context.

When rte_ckeditor is installed, all caches are cleared
and the first hit to a TYPO3 installation is a CLI command, the
backend RequireJS paths are calculated and saved wrongly
causing the backend to fail afterwards.

How to replicate:

* install rte_ckeditor
* delete var/Cache/
* run typo3/cli_dispatch.phpsh
* visit backend

Resolves: #80183
Releases: master
Change-Id: I5de9462a6f577f2eb6c06adabe44a6a86e3c2592
Reviewed-on: https://review.typo3.org/51981
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Anders Kostending <aha@systime.dk>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>