[BUGFIX] Add css for marking broken links in rte_ckeditor
The styling for broken page links that was added in
RteHtmlParser::markBrokenLinks is by default removed by ckeditor
so this functionality is no longer available by default.

Page links which link to a non-existing page currently already
get the additional attribute a[data-rte-error]. This patch adds
CSS to visually style broken links with a red border and yellow
background. This is the same styling that was previously used.

This is done by adding the plugin showbrokenlinks.

The style is still added in markBrokenLinks. These attributes
will be removed by ckeditor by default. However, the functionality
is kept to function as fallback.

Resolves: #84987
Releases: master, 8.7
Change-Id: I63e1e1db248fb2c343cdd19b41424621b684e87e
Reviewed-on: https://review.typo3.org/56943
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Sybille Peters <sypets@gmx.de>
Tested-by: Sybille Peters <sypets@gmx.de>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>