[TASK] CKEditor RTE: Add button to insert soft hyphen
This patch re-adds a missing functionality to CKEditor.

Due to the move from htmlArea to CKEditor, the well-known and actually
important button/shortcut "soft hyphen" for text in responsive webdesign
was gone.

The button is added per default to the existing CKEditor presets "default"
and "full". Existing translations from htmlArea have been migrated and
more were added.

Pasting a soft hyphen is possible via the UI button and a keyboard
shortcut: 'CTRL' + '-'. The shortcut can be deactivated in custom
RTE presets.

Resolves: #85361
Releases: master, 8.7
Change-Id: Iec97f4c9d1ede269e2df5cb4fe795df0ef5ca558
Reviewed-on: https://review.typo3.org/57373
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Wittkiel Gruppe <ts@wittkiel-gruppe.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>