[TASK] CKEditor RTE: Re-add the wordcount functionality
This patch re-adds a missing functionality to CKEditor.

Due to the move from htmlArea to CKEditor, the well-known
wordcount information was gone.

The wordcount information is added to all existing CKEditor presets
per default. In addition, it is now possible to customize the displayed
information below the editor.

- It's possible to switch paragraph-, word- and/or character count.
- You can define if spaces are counted as charactars.
- You can define if HTML tags are taken into account.
- You can define if linebreaks are taken into account.

Beside that, the used plugin enables also setting up maximum
values for word- and/or character count. This can be easily used
for custom RTE presets. The fact that the global configuration is
located in the Base.yaml makes it possible to overwrite settings
partially or completely in any inherited RTE preset. With that it is
easy possible to realize word/character limits.

Following commands were used to add the external package as
a dependency and copy the plugin files/folders to the final location:
`yarn add --dev ckeditor-wordcount-plugin && yarn build`

Resolves: #86395
Releases: master
Change-Id: I2c76d7bfebc9a5a57c1f105788af206bdffb00f2
Reviewed-on: https://review.typo3.org/58405
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>