[BUGFIX] Keep update button in linkbrowser for RTE CKEditor on edit
This patch moves the current value of the link into the parameter array
for the linkbrowser to keep the value in all linked actions inside the
linkbrowser for the RTE CKEditor.

This keeps the update button in the page linkhandler,
if an existing page link will be edited.

Resolves: #83545
Releases: master, 8.7
Change-Id: Ia9217057e45a5d94668bb5ac2f111198da9a69d1
Reviewed-on: https://review.typo3.org/58737
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>