[BUGFIX] Prevent overriding ckeditor config from plugins
Configuration from external plugins is now only set if the plugin actually
provided some. By default, the plugin name is used for any configuration
provided by an external plugin. Since the name of the internal configuration
setting can be chosen by the author of a plugin we now provide a new option
to adjust the name that should be used.

Resolves: #88302
Releases: master, 9.5, 8.7
Change-Id: I5c1f921f4897cf15e11199015454d0927915a30e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63891
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>