[BUGFIX] Mark field as changed, when changing text in ckeditor source mode
Changing content in the 'source' mode textarea of ckeditor does not
automatically trigger a change event for its parent textarea. Because
of that, the field was never marked as "changed" in the FormEngine and
the user would then not receive a warning modal when closing the form
without saving first. This behaviour is now fixed.

Resolves: #90621
Releases: master, 9.5
Change-Id: I4ecfee499ce5abda874c285a8babde29fb47c955
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63675
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>