[TASK] Use plain script-loading for CKEditor v4
CKEditor is now loaded via a plain <script async> tag in
preparation for a transition from requirejs to ES6 modules.

CKEditor v4 is not available as ES6 module and can not easily
be transformed into a strict-mode compatible module.
CKEditor v4 requires `this` to be the global window object
in various functions, but `this` is not bound to the global
window object in strict-mode (as implied by <script type="module">).

Note: CKEditor is not a real AMD module anyway, but is configured
to be shimed by requirejs to be loadable as AMD module,
therefore it can safely be loaded using good old <script>.

Releases: main
Resolves: #96394
Related: #96323
Change-Id: Ife0b476ba29b85f85dfd654ea096cdf30c47b6ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72733
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>