[BUGFIX] Correct prefixing of additional css files for ckeditor5
Some CSS minifier remove the semicolon before the curly brace
While this is valid CSS, the ScssPHP Parser is unable to identify
the end of a declaration block. We are adding them, to avoid
parsing errors. Superfluous semicolons will be dropped by the parser

Moving CSS variables assigned to :root to the new parent.

Resolves: #98605
Releases: main
Change-Id: I6e24e94ea7415cf94aa8527aedfe5a4548288705
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76109
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>