[BUGFIX] Correct editor content CSS preprocessing
The CSS preprocessing makes it easier to provide custom
content CSS for the CKEditor5. The integrator does not need
to take care of prefixing or scoping the custom CSS to avoid
collisions with the existing backend CSS.

To achieve this we are remapping body, html and :root
definitions to the individual CKEditor5 instances.

We are adjusting the processing of body and html element
selectors to avoid accidental replacements hyphenated
strings and allow combined attribute selectors.

In addition, we are now replacing multi-selectors correctly.

Example for an hyphenated string (ignored):
- font-family: var(--bs-body-font-family);

Example for an combined attribute selector (replaced):
- html[dir="rtl"]

Example for an multi selector (replaced):
- html[dir="rtl"], body

Resolves: #99445
Releases: main
Change-Id: I330859b52390d02f1aa0140e9bc6e8d3123389d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77243
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.com>