[TASK] Improve CKEditor5 Plugin handling and defaults
This is a prepatch to handle plugins and their configuration
in general. Plugin names differ between CKEditor4 and
CKEditor5, and there is no automatic migration of Plugin
names. To keep existing configuration working, we are
providing a mapping between old and new plugin names.

When removing a plugin, we also need to remove the
configuration associated to this plugin before passing
it to CKEditor. A reasonable default for the plugin is
now provided through the migration, if no configuration
is set.

Resolves: #99494
Releases: main
Change-Id: I14d00670cbf566be7666717ff7d254ee7c4c65b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77290
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>