[BUGFIX] Provide tag to add JavaScript Modules to importmap in Backend Form
The JavaScript module import map is static and only generated and
loaded in the first request to a document. All possible future
modules requested in later Ajax calls need to be registered already
in the first initial request.

We are adding a new tag 'backend.form' that is used to identify
JavaScript modules that can be used within the backend forms. This
will ensure that the import maps are available for these modules even
if the element is not displayed directly.

A typical use case for this is an InlineRelationRecord where the
CKEditor is not part of the main record but needs to be loaded for the
child record.

Resolves: #99490
Releases: main
Change-Id: Ie30e70a37abcebb0b9381b29ec9f55730a93976a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77317
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Jan Delius <me@jandelius.eu>
Tested-by: Rico Drexler <rico.drexler@gmx.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>