[TASK] Allow to configure CKEditor "General HTML Support" plugin
The CKEditor 5 build for TYPO3 contains the "General HTML Support"
plugin. It is however not possible to configure the plugin, since
the `htmlSupport` configuration key is not added to the CKEditor
configuration.

With this patch, the `htmlSupport` configuration key is passed
through to the general CKEditor configuration.

See https://ckeditor.com/docs/ckeditor5/latest/features/html/general-html-support.html

Resolves: #100861
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ieb4a28bf777632aa2d7e2c68a884f8e7a0174509
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79015
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>