[BUGFIX] Allow attributes for anchors in ckeditor
Ckeditor 5 is used since TYPO3 12 which has significantly changed
how elements in the DOM are handled.

The `a` element is handled in the plugin typo3-link, but it did not
handle some attributes correctly, specifically:

- class
- rel
- target

This resulted in only `href` and `title` being persisted when saving
or switching between source and normal mode.

This is now fixed, so that all 5 attributes will be handled
correctly.

Resolves: #101360
Related: #96874
Releases: main, 12.4
Change-Id: I403ec880a24ae582a8c73b2ac2ccae49ffcbc21f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80131
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>