[BUGFIX] Ensure CKEditor ui language is always set
If the CKEditor configuration has a language configuration, but does not
define the `ui` language setting, no fallback language is added to the
editor configuration and an undefined array key exception is thrown.

This change ensures, that the `ui` language is properly set when
the editor configuration has a language configuration and the
`ui` language is missing.

Resolves: #101401
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Id8bc03fd832413cf6ed1f9c5d443fa9f3387811c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80112
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>