[BUGFIX] Properly evaluate selected class in BrowseLinksController
Generally, adding a class to a link in CKeditor is done
via the corresponding link browser (BrowseLinksController),
which renders a select field with a list of allowed classes
(as configured). However, CKeditor  also automatically applies
a so called "highlighting class" to a link. Furthermore
it's always possible to manually set additional classes
via the "Source" view mode. To be able to properly resolve
an actually selected (allowed) class for the select field,
the current link class value is now cleanup up before
markup generation takes place in the BrowseLinksController.

Resolves: #101436
Releases: main, 12.4
Change-Id: Ibed202dfe86b94524d9f2d6a5eb7587bef8a21f1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80215
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>