[BUGFIX] Fix link attribute handling in ckeditor
Fix handling link attributes (such as target, title)
in link browser. The attributes were formerly not
always applied to the link currently being edited.

The problem appeared to be that the current selection
in the editor lost focus and when link attributes
were handled for this case, only the href attribute
was applied.

This is a follow-up patch for patch
https://review.typo3.org/c/Packages/TYPO3.CMS/+/80034
which did not fully resolve problems with setting the
link attributes.

Resolves: #101638
Related: #101360
Related: #101205
Releases: main, 12.4
Change-Id: Id5aeaae11341169b7a7bda15f62aeb2937486f88
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80528
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Benjamin Franzke <ben@bnf.dev>