[TASK] Improve CKEditor vendor logo placement
With CKEditor v38.0.0 a small "Powered by CKEditor" logo has been added
to the bottom right of the editor component. The logo is visible as soon
as the CKEditor component receives the focus. The current placement of
the logo does however overlay the border of the editor component and the
term "Powered by" does not suit well to the TYPO3 backend.

With this change, the CKEditor configuration has been extended, so it is
now possible to configure `config.ui.*` settings. The `Base.yaml` file
has been extended with CKEditor UI specific settings, which remove the
term "Powered by" from the logo and which optimize the logo placement,
so it appears in the bottom right corner of the editor component without
overlaying the border.

Resolves: #101759
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ic35054a92c7101fac9381e1476c214ffad3e0465
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80678
Tested-by: Jasmina Lie�mann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Lie�mann <minapokhalo+typo3@gmail.com>