[BUGFIX] Allow multiple CSS classes in link browser again
When a link is selected in CKEditor 5, the class `ck-link_selected` is
added to the link. This caused the link browser to not show the selected class, which has been fixed with #101436. The bugfix does however not consider, that a CSS class in link browser can consist of a string with multiple classes separated by a space.

This change ensures, that the link browser again accepts a class string
containing multiple class names and shows the selected class string in
the CSS-Class select field.

Resolves: #101959
Related: #101436
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I278df61e035073a54ba176955141593c7cd9f3cf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81145
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>