[BUGFIX] Add CKEditor5 migrator for {extraA,disa,a}llowedContent
Allowed content in CKEditor5 is to be configured via the
General HTML Support plugin option config.htmlSupport.
CKEditor4 style configs are now parsed and migrated.

Note that this config.htmlSupport only applies to elements that
are "known" to CKEditor5. Tags like "svg" or custom elements like
"<foo-bar>" are not configurable this way as `htmlSupport.(dis)allow`
can only handle elements that are defined in the CKEditor5 schema. [1]

Also reintroduce the default configuration that allows any class
and any attribute that starts with data- to be specified/preserved on
known tags. This was removed with #96874.

[1] https://ckeditor.com/docs/ckeditor5/latest/features/html/general-html-support.html#enabling-custom-elements

Resolves: #99738
Releases: main, 12.4
Change-Id: Idaadfc09328efa4d1f21171a400dc45d29ed5904
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81316
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>