[BUGFIX] Enable configuration passthrough for custom CKEditor5 plugins
Use JavaScript object destructuring to "remove" all options from the
configuration object, that are consumed by our CKEditor5 wrapper.
Also cleanup RTE resource and config handling to not write
unused options into the configuration array and to avoid resolving
legacy (CKEditor4) resource paths which are dropped anyway.

As a drive-by, dots are now substituted with '_' in RTE field IDs to
avoid the following hassle:
Dots are interpreted as CSS classes when the ID value is used
in combination with a number sign (#) to create a CSS selector
for the respective field ID. That means the selector will not match.
The class selector additionally becomes invalid once there is a digit
after the dot, as CSS classes need to start with strings.
(Example: EXT:styleguide in_flex » tab » rte.2).

Resolves: #100784
Resolves: #101437
Releases: main, 12.4
Change-Id: I076b838c03588ad6eb8ad075a9df58501f146376
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81406
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>