[DOCS] Add details for CSS-prefixing in EXT:rte_ckeditor
The current automatism of auto-prefixing might not be
easily understandable, so the process is now described
more clearly, with an example.

Also `:root` is mentioned to apply additional reset
possibilites.

A note that the CSS file is cached (for now) and has
no automatic cache busting is added, too.

Resolves: #105329
Releases: main, 12.4
Change-Id: I4b0b570d2327cead90e618423d18cb5975f0fe54
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86614
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>