[!!!][TASK] Set default light/dark context awareness for CKEditor
With #105640, context awareness for CKEditor RTE was introduced, but
not enabled by default to not break existing customizations/enhancements
of the editor interface.

The TYPO3 Core default stylesheet [1] is now adapted to enable the
context awareness by default. Prior white backgrounds are now shown
as either dark or light backgrounds, depending on the editor's
preference. Color assignments of "links" are slightly adjusted,
Other existing color assignments (broken links, blockquotes) are fine
in both modes.

Please note that the display of RTE contents in the frontend is
unaffected.

[1] EXT:rte_ckeditor/Resources/Public/Css/contents.css

Resolves: #106964
Related: #105640
Releases: main
Change-Id: I046e181671dca9257da77564c2d533c816bac983
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89862
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>